feat: chunk bmt functions and inclusion proof
